requirements.txt
absl-py==0.7.1
astor==0.8.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorama==0.4.1
cycler==0.10.0
dash==1.0.2
dash-bootstrap-components==0.6.1
dash-core-components==1.0.0
dash-html-components==1.0.0
dash-renderer==1.0.0
dash-table==4.0.2
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.1
Flask-Compress==1.4.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
idna==2.6
imageio==2.5.0
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
itsdangerous==1.1.0
jedi==0.14.1
Jinja2==2.10.1
jsonschema==3.0.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
loremipsum==1.0.5
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
networkx==2.3
numpy==1.16.3
opencv-contrib-python==4.1.0.25
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
Pillow==6.1.0
plotly==2.7.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.0
Pygments==2.4.2
pyparsing==2.4.0
pyrsistent==0.15.3
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
pywinpty==0.5.5
PyYAML==5.1.1
pyzmq==18.0.2
qtconsole==4.5.1
requests==2.18.4
scikit-image==0.15.0
scipy==1.3.0
Send2Trash==1.5.0
six==1.12.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.5.0
wincertstore==0.2
wrapt==1.11.2
